scryptenc: add SCRYPT_* exit codes
This will be a multi-commit process, with one commit per replacement (to
make it easier to verify).

Strictly speaking, we don't need to #include "scryptenc.h" in
scryptenc_cpuperf.c yet.  But I'm adding it now so that the next few
commits will be easier to verify (i.e. no non-replacement changes).

Normally we'd require these #defines to begin with SCRYPTENC_.  However,
that would make the line lengths too long and/or awkward.  Namely:
- We definitely want error #5 to include "_EOPENSSL" in the #define.
- To avoid rewriting the error explanations, the #define names must be
  less than 2 tabs-worth (so 15 characters).
- 15 chars minus strlen("_EOPENSSL") is 6, which fits "SCRYPT" exactly.